fsp1.1: Mark graphics init done after SiliconInit phase
If the VBT was provided to the FSP GOP driver then graphics init
will be done as part of SiliconInit step and we can mark that
when it is completed.

This will result in the "oprom" flag being set properly in the
coreboot gpio table and the netboot firmware will have video.

BUG=chrome-os-partner:50864
BRANCH=glados
TEST=boot image.net.bin on chell and get working graphics
without being setuck in a reboot loop thinking graphics needs
to be started when it already has been.

Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/331301
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
(cherry picked from commit eeb9d470d8118422feb39ca71106972f2882e240)
Change-Id: Ic59bad27eb9f184ca3eba24643851bfadfe23ab5
Reviewed-on: https://chromium-review.googlesource.com/331355