Cyan: Fix Touchscreen Interrupt
Elan touchscreen driver expects the first gpio resource in asl
to be the reset line.
The driver considers the gpio based irq line as reset gpio resource
and changes the direction to output.
This will cause irq registration to fail.

Solution is to pass Interrupt resource for touchscreen irq
instead of GpioInt.

BRANCH=firmware-cyan-7287.57.B
BUG=chrome-os-partner:51154
TEST=verify touchscreen functionality

Change-Id: Ia72d4ad80117f3c0014098113c9027416026e65e
Signed-off-by: Jagadish Krishnamoorthy <jagadish.krishnamoorthy@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/334622
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>