skylake: Only apply DdrFreqLimit if set in devicetree.cb
This commit changes the FSP UPD value setup for DdrFreqLimit to only
be applied from devicetree.cb if it is actually set in there.

This allows mainboard code to override DdrFreqLimit for a single
memory config while still leaving it flexible enough to override
for all memory using devicetree.

BUG=chrome-os-partner:51209
BRANCH=glados
TEST=emerge-chell coreboot

Change-Id: I8e7e1941c6054b81f975e808e181a219b7da738a
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/336307
Reviewed-by: Aaron Durbin <adurbin@chromium.org>