Lars: Add sdmode-delay device property for maxim98357a
Add "sdmode-delay" as a device property for the maxim98357a codec.
This speaker amp requires both SFRM & BCLK to active and stable
before it is unmuted. If there is a BCLK and no SFRM,
that results in a pop noise.
Adding a configurable delay parameter for all Skylake platforms
to allow sufficient time for the BCLK & SFRM on I2S to be stable
before the amp unmutes itself to avoid a pop noise at the start of playback.
Setting the delay to 5ms since the observed delay between SFRM and unmuting
of the amp is around 2ms.

BUG=chrome-os-partner:50632
BRANCH=firmware-glados-7820.B
TEST=boot to the kernel & read the device property in maxim codec.

Change-Id: I1fff4f86ff816e907553e7a6f1d05713f9d85084
Signed-off-by: Rohit Ainapure <rohit.m.ainapure@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/337750
Reviewed-by: Aaron Durbin <adurbin@chromium.org>