checkpatch: Ignore LINE_SPACING
The newest checkpatch version now flags code such as

    int i;
    for (i = 0; ...

because there is no blank line between a declaration and the next
statement. We have always allowed variable declarations inline with the
code in coreboot and libpayload, so this doesn't make much sense for
us. Add it to the ignore list.

BRANCH=None
BUG=None
TEST=None

Change-Id: I865d69ccdd30e89083cfee5ca9c2332bfc6c1970
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/341513
Reviewed-by: Aaron Durbin <adurbin@chromium.org>