rockchip/rk3399: Actually remove big CPU initialization from bootblock
CL:377541 was supposed to remove the big CPU cluster initialization from
rkclk_init() in the bootblock and move it to a more suitable place in
ramstage. Except that next to all the code cleanup I did in that patch,
I seem to have forgotten to actually remove that old code.

Big thanks to Nico for spotting that in the upstream coreboot review.

BRANCH=gru
BUG=chrome-os-partner:54906
TEST=Booted Kevin.

Change-Id: I13dab208225b7e43ad864f2f3cf51b3c104acd4b
Reported-by: Nico Huber <nico.h@gmx.de>
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/389253