Umaro: Update DPTF policy for TRS1 and parameter for TRS2
As the TRS1 is placed near the charger IC on Umaro board, The
original DPTF policy TRS1 impact SOC power so much, which is not
nessary. so this change make TRS1 control the charger current only.

The TRS2 is placed near DDR on Umaro board. So the new parameter
won't make it impact the SOC power so much, which also meet thermal
test requirment.

BUG=chrome-os-partner:60221
BRANCH=firmware-strago-7287.B
TEST=build and boot on umaro dut

Change-Id: I33ead6d11bfd62335a37e1ef6f65df11224463ec
Signed-off-by: Guohua Zhong <ghzhong@yifangdigital.com>
Reviewed-on: https://chromium-review.googlesource.com/413637
Reviewed-by: Aaron Durbin <adurbin@chromium.org>